requirements.txt
fastapi
gradio
gradio_pdf
boto3
google-generativeai

# Utilities
uvicorn[standard]
python-dotenv
itsdangerous
# selectolax
# pdf2image
